backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.1.0
tiktoken==0.5.2
sentence-transformers>=2.6.0
elasticsearch==8.11.0
python-dotenv==1.0.0
ollama==0.3.3